// mobile-expo/graphql/gql/generated-models.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** Cursor for paging through collections */
  ConnectionCursor: { input: any; output: any; }
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export type CursorPaging = {
  /** Paginate after opaque cursor */
  after?: InputMaybe<Scalars['ConnectionCursor']['input']>;
  /** Paginate before opaque cursor */
  before?: InputMaybe<Scalars['ConnectionCursor']['input']>;
  /** Paginate first */
  first?: InputMaybe<Scalars['Int']['input']>;
  /** Paginate last */
  last?: InputMaybe<Scalars['Int']['input']>;
};
// ...
export type DateFieldComparison = {
  between?: InputMaybe<DateFieldComparisonBetween>;
  eq?: InputMaybe<Scalars['DateTime']['input']>;
  gt?: InputMaybe<Scalars['DateTime']['input']>;
  gte?: InputMaybe<Scalars['DateTime']['input']>;
  in?: InputMaybe<Array<Scalars['DateTime']['input']>>;
  is?: InputMaybe<Scalars['Boolean']['input']>;
  isNot?: InputMaybe<Scalars['Boolean']['input']>;
  lt?: InputMaybe<Scalars['DateTime']['input']>;
  lte?: InputMaybe<Scalars['DateTime']['input']>;
  neq?: InputMaybe<Scalars['DateTime']['input']>;
  notBetween?: InputMaybe<DateFieldComparisonBetween>;
  notIn?: InputMaybe<Array<Scalars['DateTime']['input']>>;
};
// ...
export type IdFilterComparison = {
  eq?: InputMaybe<Scalars['ID']['input']>;
  gt?: InputMaybe<Scalars['ID']['input']>;
  gte?: InputMaybe<Scalars['ID']['input']>;
  iLike?: InputMaybe<Scalars['ID']['input']>;
  in?: InputMaybe<Array<Scalars['ID']['input']>>;
  is?: InputMaybe<Scalars['Boolean']['input']>;
  isNot?: InputMaybe<Scalars['Boolean']['input']>;
  like?: InputMaybe<Scalars['ID']['input']>;
  lt?: InputMaybe<Scalars['ID']['input']>;
  lte?: InputMaybe<Scalars['ID']['input']>;
  neq?: InputMaybe<Scalars['ID']['input']>;
  notILike?: InputMaybe<Scalars['ID']['input']>;
  notIn?: InputMaybe<Array<Scalars['ID']['input']>>;
  notLike?: InputMaybe<Scalars['ID']['input']>;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** The cursor of the last returned record. */
  endCursor?: Maybe<Scalars['ConnectionCursor']['output']>;
  /** true if paging forward and there are more records. */
  hasNextPage?: Maybe<Scalars['Boolean']['output']>;
  /** true if paging backwards and there are more records. */
  hasPreviousPage?: Maybe<Scalars['Boolean']['output']>;
  /** The cursor of the first returned record. */
  startCursor?: Maybe<Scalars['ConnectionCursor']['output']>;
};
// ...
export type QueryServiceLogsArgs = {
  filter?: ServiceLogFilter;
  paging?: CursorPaging;
  sorting?: Array<ServiceLogSort>;
};
// ...
export type ServiceLog = {
  __typename?: 'ServiceLog';
  bills?: Maybe<ServiceLogBillsConnection>;
  billsAggregate: Array<ServiceLogBillsAggregateResponse>;
  createdAt: Scalars['DateTime']['output'];
  customGarage?: Maybe<Scalars['String']['output']>;
  date?: Maybe<Scalars['DateTime']['output']>;
  deletedAt?: Maybe<Scalars['DateTime']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  garage?: Maybe<Garage>;
  id: Scalars['ID']['output'];
  links?: Maybe<Array<Scalars['String']['output']>>;
  media?: Maybe<Array<Scalars['String']['output']>>;
  mileage?: Maybe<Scalars['Float']['output']>;
  type: ServiceLogType;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  vehicle: UserVehicle;
};
// ...
export type ServiceLogConnection = {
  __typename?: 'ServiceLogConnection';
  /** Array of edges. */
  edges: Array<ServiceLogEdge>;
  /** Paging information */
  pageInfo: PageInfo;
  /** Fetch total count of records */
  totalCount: Scalars['Int']['output'];
};
// ...
export type ServiceLogEdge = {
  __typename?: 'ServiceLogEdge';
  /** Cursor for this node. */
  cursor: Scalars['ConnectionCursor']['output'];
  /** The node containing the ServiceLog */
  node: ServiceLog;
};

export type ServiceLogFilter = {
  and?: InputMaybe<Array<ServiceLogFilter>>;
  bills?: InputMaybe<ServiceLogFilterServiceLogBillFilter>;
  createdAt?: InputMaybe<DateFieldComparison>;
  date?: InputMaybe<DateFieldComparison>;
  deletedAt?: InputMaybe<DateFieldComparison>;
  description?: InputMaybe<StringFieldComparison>;
  garage?: InputMaybe<ServiceLogFilterGarageFilter>;
  id?: InputMaybe<IdFilterComparison>;
  or?: InputMaybe<Array<ServiceLogFilter>>;
  type?: InputMaybe<ServiceLogTypeFilterComparison>;
  updatedAt?: InputMaybe<DateFieldComparison>;
  vehicle?: InputMaybe<ServiceLogFilterUserVehicleFilter>;
};

export type ServiceLogFilterGarageFilter = {
  and?: InputMaybe<Array<ServiceLogFilterGarageFilter>>;
  createdAt?: InputMaybe<DateFieldComparison>;
  deletedAt?: InputMaybe<DateFieldComparison>;
  gplace_id?: InputMaybe<StringFieldComparison>;
  id?: InputMaybe<IdFilterComparison>;
  lat?: InputMaybe<NumberFieldComparison>;
  lng?: InputMaybe<NumberFieldComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<ServiceLogFilterGarageFilter>>;
  phone_number?: InputMaybe<StringFieldComparison>;
  status?: InputMaybe<GarageStatusFilterComparison>;
  updatedAt?: InputMaybe<DateFieldComparison>;
};

export type ServiceLogFilterServiceLogBillFilter = {
  and?: InputMaybe<Array<ServiceLogFilterServiceLogBillFilter>>;
  createdAt?: InputMaybe<DateFieldComparison>;
  deletedAt?: InputMaybe<DateFieldComparison>;
  id?: InputMaybe<IdFilterComparison>;
  or?: InputMaybe<Array<ServiceLogFilterServiceLogBillFilter>>;
  total?: InputMaybe<NumberFieldComparison>;
  updatedAt?: InputMaybe<DateFieldComparison>;
};

export type ServiceLogFilterUserVehicleFilter = {
  and?: InputMaybe<Array<ServiceLogFilterUserVehicleFilter>>;
  createdAt?: InputMaybe<DateFieldComparison>;
  customModel?: InputMaybe<StringFieldComparison>;
  deletedAt?: InputMaybe<DateFieldComparison>;
  id?: InputMaybe<IdFilterComparison>;
  name?: InputMaybe<StringFieldComparison>;
  or?: InputMaybe<Array<ServiceLogFilterUserVehicleFilter>>;
  status?: InputMaybe<UserVechileStatusFilterComparison>;
  updatedAt?: InputMaybe<DateFieldComparison>;
};
// ...
export type ServiceLogSort = {
  direction: SortDirection;
  field: ServiceLogSortFields;
  nulls?: InputMaybe<SortNulls>;
};

export enum ServiceLogSortFields {
  CreatedAt = 'createdAt',
  Date = 'date',
  DeletedAt = 'deletedAt',
  Description = 'description',
  Id = 'id',
  Type = 'type',
  UpdatedAt = 'updatedAt'
}

export type ServiceLogSumAggregate = {
  __typename?: 'ServiceLogSumAggregate';
  type?: Maybe<Scalars['Float']['output']>;
};

export enum ServiceLogType {
  Detailing = 'DETAILING',
  Maintenance = 'MAINTENANCE',
  Other = 'OTHER',
  Retrofit = 'RETROFIT',
  Tuning = 'TUNING',
  Upgrading = 'UPGRADING'
}

export type ServiceLogTypeFilterComparison = {
  eq?: InputMaybe<ServiceLogType>;
  gt?: InputMaybe<ServiceLogType>;
  gte?: InputMaybe<ServiceLogType>;
  iLike?: InputMaybe<ServiceLogType>;
  in?: InputMaybe<Array<ServiceLogType>>;
  is?: InputMaybe<Scalars['Boolean']['input']>;
  isNot?: InputMaybe<Scalars['Boolean']['input']>;
  like?: InputMaybe<ServiceLogType>;
  lt?: InputMaybe<ServiceLogType>;
  lte?: InputMaybe<ServiceLogType>;
  neq?: InputMaybe<ServiceLogType>;
  notILike?: InputMaybe<ServiceLogType>;
  notIn?: InputMaybe<Array<ServiceLogType>>;
  notLike?: InputMaybe<ServiceLogType>;
};
// ...
export enum SortDirection {
  Asc = 'ASC',
  Desc = 'DESC'
}

/** Sort Nulls Options */
export enum SortNulls {
  NullsFirst = 'NULLS_FIRST',
  NullsLast = 'NULLS_LAST'
}

export type StringFieldComparison = {
  eq?: InputMaybe<Scalars['String']['input']>;
  gt?: InputMaybe<Scalars['String']['input']>;
  gte?: InputMaybe<Scalars['String']['input']>;
  iLike?: InputMaybe<Scalars['String']['input']>;
  in?: InputMaybe<Array<Scalars['String']['input']>>;
  is?: InputMaybe<Scalars['Boolean']['input']>;
  isNot?: InputMaybe<Scalars['Boolean']['input']>;
  like?: InputMaybe<Scalars['String']['input']>;
  lt?: InputMaybe<Scalars['String']['input']>;
  lte?: InputMaybe<Scalars['String']['input']>;
  neq?: InputMaybe<Scalars['String']['input']>;
  notILike?: InputMaybe<Scalars['String']['input']>;
  notIn?: InputMaybe<Array<Scalars['String']['input']>>;
  notLike?: InputMaybe<Scalars['String']['input']>;
};